organize graphql code by role
